+++ colorpicker/src/jvmTest/kotlin/codes/side/colorpicker/ui/ChannelPlaneTest.kt
@@ -45,6 +45,7 @@ import codes.side.color.Okhsl
 import codes.side.color.Okhsv
 import codes.side.color.Srgb
 import codes.side.color.compose.toComposeColor
+import codes.side.colorpicker.foundation.requirePlaneChannels
 import codes.side.colorpicker.state.ColorPickerState
 import codes.side.colorpicker.state.assertNear
 import kotlin.math.abs